♻️ [Refactor]:: Curriculum 컴포넌트 리팩토링 및 최적화
